Add collapsible Date filter
