+++ src/components/index.js
@@ -1,10 +1,10 @@
-export { default as SearchkitProvider } from "./SearchkitProvider";
-export { default as HitsTable } from "./HitsTable";
-export { default as Pagination } from "./Pagination";
-export { default as DateFilter } from "./DateFilter";
-export { default as ActiveFiltersDisplay } from "./ActiveFiltersDisplay";
-export { default as ClearFiltersDisplay } from "./ClearFiltersDisplay";
-export { default as MentionsFilter } from "./MentionsFilter";
-export { default as Sorter } from "./Sorter";
-export { default as ExposeQuery } from "./ExposeQuery";
-export * from "./Search";
+export { default as SearchkitProvider } from './SearchkitProvider';
+export { default as HitsTable } from './HitsTable';
+export { default as Pagination } from './Pagination';
+export { default as ActiveFiltersDisplay } from './ActiveFiltersDisplay';
+export { default as ClearFiltersDisplay } from './ClearFiltersDisplay';
+export { default as MentionsFilter } from './MentionsFilter';
+export { default as Sorter } from './Sorter';
+export { default as ExposeQuery } from './ExposeQuery';
+export * from './DateFilter';
+export * from './Search';
